style: run Prettier on all the Solidity files

+++ contracts/interfaces/callback/IPrimitiveLendingCallback.sol
@@ -4,6 +4,12 @@ pragma solidity 0.8.0;
 import "../../libraries/Position.sol";
 
 interface IPrimitiveLendingCallback {
-    function borrowCallback(uint deltaL, uint deltaRisky, uint deltaStable, bytes calldata data) external;
-    function repayFromExternalCallback(uint deltaStable, bytes calldata data) external;
+    function borrowCallback(
+        uint256 deltaL,
+        uint256 deltaRisky,
+        uint256 deltaStable,
+        bytes calldata data
+    ) external;
+
+    function repayFromExternalCallback(uint256 deltaStable, bytes calldata data) external;
 }
