+++ contracts/test/engine/EngineRemove.sol
@@ -5,7 +5,6 @@ import "../../interfaces/IPrimitiveEngine.sol";
 import "../../interfaces/IERC20.sol";
 
 contract EngineRemove {
-
     address public engine;
     address public risky;
     address public stable;
@@ -13,30 +12,45 @@ contract EngineRemove {
 
     constructor() {}
 
-    function initialize(address _engine, address _risky, address _stable) public {
-      engine = _engine;
-      risky = _risky;
-      stable = _stable;
+    function initialize(
+        address _engine,
+        address _risky,
+        address _stable
+    ) public {
+        engine = _engine;
+        risky = _risky;
+        stable = _stable;
     }
 
-    function removeToMargin(bytes32 pid, uint dLiquidity, bytes memory data) public  {
+    function removeToMargin(
+        bytes32 pid,
+        uint256 dLiquidity,
+        bytes memory data
+    ) public {
         IPrimitiveEngine(engine).remove(pid, dLiquidity, true, data);
     }
 
-    function removeToExternal(bytes32 pid, uint dLiquidity, bytes memory data) public  {
+    function removeToExternal(
+        bytes32 pid,
+        uint256 dLiquidity,
+        bytes memory data
+    ) public {
         IPrimitiveEngine(engine).remove(pid, dLiquidity, false, data);
     }
 
-    function removeCallback(uint dRisky, uint dStable, bytes memory data) public  {
-      return;
+    function removeCallback(
+        uint256 dRisky,
+        uint256 dStable,
+        bytes memory data
+    ) public {
+        return;
     }
 
-    function getPosition(bytes32 pid) public view returns(bytes32 posid) {
+    function getPosition(bytes32 pid) public view returns (bytes32 posid) {
         posid = keccak256(abi.encodePacked(address(this), pid));
     }
 
     function name() public view returns (string memory) {
-      return "EngineRemove";
+        return "EngineRemove";
     }
 }
-
